+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-08-11 10:11:43
-// Commit: style(state): adjust design tokens, typography, and contrast in state
+// Updated at: 2026-08-11 10:19:29
+// Commit: feat(state): integrate offline caching for state state
 
 class stateComponentState {
-  final String commitRef = '3514';
-  final String lastMessage = 'style(state): adjust design tokens, typography, and contrast in state';
-  final DateTime timestamp = DateTime.parse('2026-08-11T10:11:43+05:00');
+  final String commitRef = '3516';
+  final String lastMessage = 'feat(state): integrate offline caching for state state';
+  final DateTime timestamp = DateTime.parse('2026-08-11T10:19:29+05:00');
 }